==> notebooks/01_RNASeq_fastqc.ipynb <==
library(stringr)
library(parallel)

samples = readLines("../data/RNAseq/sample_list")

samples = readLines("../data/RNAseq/sample_list_2")

home = '/nfs/lab/projects/islet_cytok/pipeline/RNAseq/sample'

fastqc = function (samp) {
outdir = paste( home, samp , "qc", sep="/")
c1     = paste( "mkdir -p",  outdir)
c2     = paste( "cd"      ,  outdir)

r1     = paste('/nfs/lab/projects/islet_cytok/data/RNAseq/', samp, '_R1_001.fastq.gz', sep="")
r2     = paste('/nfs/lab/projects/islet_cytok/data/RNAseq/', samp, '_R2_001.fastq.gz', sep="")

c3     = paste( "fastqc --outdir", outdir,  '--nogroup --threads 2' , r1, r2)

system(c1)
system(c2)
system(c3)
    }

invisible(mclapply(samples, fastqc, mc.cores = 32))

### New batch

fi = list.files('/nfs/lab/projects/islet_dex/data/RNAseq/RNAseq_supp/')
fi = fi[grepl("islet_dex", fi)]
samples = str_split_fixed(fi, '_R',2)[,1]

for (fl in fi){
    system(
    paste0('ln -s /nfs/lab/projects/islet_dex/data/RNAseq/RNAseq_supp/', 
           fl, " /nfs/lab/projects/islet_cytok/data/RNAseq/")
    )
}

invisible(mclapply(samples, fastqc, mc.cores = 32))
