==> notebooks/03_RNASeq_expressionRSEM.ipynb <==
library(stringr)
library(parallel)

samples = readLines("../data/RNAseq/sample_list")
samples = readLines("../data/RNAseq/sample_list_2")
home = '/nfs/lab/projects/islet_cytok/pipeline/RNAseq/sample'

samples =gsub("L001", "combined", samples)

rsem = function(samp){
outdir = paste( home, samp , "rsem", sep="/")
c1     = paste( "mkdir -p",  outdir)
c2     = paste( "cd"      ,  outdir)

input_bam = paste(home, "/", samp, '/alignment/', samp, '_Aligned.toTranscriptome.out.bam', sep="")
    
c3 = paste(
"rsem-calculate-expression --bam --paired-end --num-threads 8 --no-bam-output --seed 3272015",
    "--estimate-rspd --forward-prob 0" ,
     input_bam,
    "/nfs/lab/publicdata/rsem_reference_hg19_20151104/rsem_reference_hg19_20151104" ,
     samp)

system(c1)
system(c2)
system(c3)
}

mclapply(samples, rsem, mc.cores = 32)

for (samp in samples){
    
 outdir = paste( home, samp , "rsem", sep="/")
   
    cm = paste ("mv ", samp, "* ", outdir, sep="" )
    system(cm)
}
