==> notebooks/05_BulkATAC_Motifs_and_figures.ipynb <==
library(stringr)
library(RColorBrewer)
library(parallel)
library(pheatmap)

setwd("../analysis/")

dir = '/nfs/lab/projects/islet_cytok/analysis/deseq_ATAC_NEW/'

fi = list.files(dir)

li= list()
for (ff in fi ){
    
    file = read.table(paste0(dir, ff), header=T, row.names = 1)
    
    nm = paste( str_split_fixed(ff, "\\.",2)[1], "down", sep=".")
     li[[nm]] <- na.omit(rownames(file)[file$padj<0.1 & file$log2FoldChange>0])## the orientation was swapped

    nm = paste( str_split_fixed(ff, "\\.",2)[1], "up", sep=".")
    li[[nm]] <- na.omit(rownames(file)[file$padj<0.1 & file$log2FoldChange<0])## the orientation was swapped
}

mm = matrix(lapply(li, length), ncol = 2, byrow = T)[c(1:4,6,5,7),1:2]
colnames(mm)= c("Down", "Up") 

##union =c(length(unique(unlist(li[c(1,3,5,7)]))) , length(unique(unlist(li[c(2,4,6,8)])))) why only the 4?
union =c(length(unique(unlist(li[c(1,3,5,7,9,11,13)]))) , length(unique(unlist(li[c(2,4,6,8,10,12,14)]))))

mm= rbind(union, mm)

pdf("cytokine_figures/Summary_DACs_cytok.pdf")
par(mfrow=c(2,2),mar = c(4,4,4,0))
b = barplot(t(mm), las=2, legend.text = colnames(mm), ylim=c(0,50000), col = brewer.pal(9, "PuOr")[c(3,7)] )
text(x = b, y =mm[,1], labels = paste0(round(unlist(mm[,1])/1000,1),"k") , pos = 1)
text(x = b, y = as.numeric(mm[,2])+ as.numeric(mm[,1]), labels = paste0(round(unlist(mm[,2])/1000,1),"k") , pos = 1)
dev.off()

u =  data.frame(unique(unlist(li)))
for(n in 1:length(li)){
    u[,n+1] =   u[,1] %in% li[[n]]
}


rownames(u)=u[,1]
u=u[,-1]

colnames(u) = names(li)

colSums(u)

bed_full = cbind(str_split_fixed(rownames(u), "\\-",4)[,2:4], rownames(u))

write.table(bed_full, "deseq_ATAC_NEW_bedfiles/All_differential.bed", sep="\t", quote=F, row.names=F)

for (cc in colnames(u)){
    bed = subset(bed_full, u[,cc]==TRUE)
    #write.table(bed, paste0("deseq_ATAC_NEW_bedfiles/", cc, ".bed"), sep="\t", quote=F, row.names=F, col.names=F)
    write.table(bed[,1:3], paste0("/nfs/lab/projects/islet_cytok/analysis/ldsc/", cc, ".bed"), sep="\t", quote=F, row.names=F, col.names=F)
}

dim(bed_full)

dim(file)

write.table( subset(bed_full, rowSums(u[,c(2,4,6,8,10,12,14)])>0),"deseq_ATAC_NEW_bedfiles/All_differential_up.bed", sep="\t", quote=F, row.names=F, col.names=F)
write.table( subset(bed_full, rowSums(u[,c(1,3,5,7,9,11,13)])>0),"deseq_ATAC_NEW_bedfiles/All_differential_down.bed", sep="\t", quote=F, row.names=F, col.names=F)

write.table( subset(bed_full[,1:3], rowSums(u[,c(2,4,6,8,10,12,14)])>0),"/nfs/lab/projects/islet_cytok/analysis/ldsc/All_differential_up.bed", sep="\t", quote=F, row.names=F, col.names=F)
write.table( subset(bed_full[,1:3], rowSums(u[,c(1,3,5,7,9,11,13)])>0),"/nfs/lab/projects/islet_cytok/analysis/ldsc/All_differential_down.bed", sep="\t", quote=F, row.names=F, col.names=F)

length(unique(rownames(u)))

write.table(u, "deseq_ATAC_NEW_bedfiles/annotated_peaks.tsv", sep="\t", quote=F)

#annotatePeaks.pl TNFa_24h_hi.bed hg19  -m known41.motif -noann > TNFa_24h_hi.NFKB_1.txt

#findMotifsGenome.pl TNFa_24h_hi.bed hg19 out -find known41.motif > TNFa_24h_hi.NFKB_v2.txt

#findMotifsGenome.pl 24h_hi.bed hg19 out -find known41.motif > 24h_hi.bed_nfkb.txt

#annotatePeaks.pl 24h_hi.bed hg19 -m known41.motif -noann > 24h_hi.bed_nfkb_v1.txt

### Plot Motif enrichemnt

homer_dir = '/nfs/lab/elisha/islets/cytokine/HOMER/homer_dec15/'
#list.files(homer_dir)
files2 = c('union_up','union_down', 'hiVhi2_up','hiVhi2_down','hi2_24_up','hi2_24_down')

#pdf("cytokine_figures/Plot_motifs.pdf", height = 6, width=12)
pdf("cytokine_figures/Plot_motifs_2.pdf")

#par(mfcol=c(2,4), mar = c(4,4,4,0))
par(mfcol=c(2,3), mar = c(4,4,4,0))
for (exp in files2){
hom= read.table( paste0(homer_dir, exp, '/knownResults.txt'), header=F, fill=T, check.names=F, skip=1)
hom$fold_change = as.numeric(gsub('%',"", hom$V7)) /(as.numeric(gsub('%',"", hom$V9))+0.1) ##pseudocount for inf (only V2)
hom$Motif = toupper(str_split_fixed(hom$V1, "\\(",5)[,1])
hom[,"-logP"] = -(hom$V4)

hom=hom[order(hom$fold_change, decreasing=T),]
hom=hom[order(hom[,'-logP'], decreasing=T),]
    
plot(hom[,c('fold_change', '-logP')], pch = 19, 
     col = (hom$V5<0.05)+1  )
if(sum(hom$V5<0.05)>0){
    text (hom[1:25,c('fold_change', '-logP')], hom$Motif[1:20], 
      pos=2, cex=0.7,offset = 0.2)
    
    #text (hom[hom$V5<0.01,c('fold_change', '-logP')], hom$Motif[hom$V5<0.01], 
    #  pos=c(2,4), cex=0.9,offset = 0.5)
}
mtext(exp)
}
dev.off()

### Run HOMER

lf    = list.files("deseq_ATAC_NEW_bedfiles")
ups   = lf[grepl("up",lf) & !grepl("All",lf)]
downs = lf[grepl("down",lf) & !grepl("All",lf)]
bup   = lf[grepl("up",lf) & grepl("All",lf)]
bdown = lf[grepl("down",lf) & grepl("All",lf)]

beddir ="/nfs/lab/projects/islet_cytok/analysis/deseq_ATAC_NEW_bedfiles"

hdir = "/nfs/lab/projects/islet_cytok/analysis/HOMER_diff_sites/"

run_homer = function(bedfile, outdir,bgfile )
{
    system(paste ('findMotifsGenome.pl', bedfile, '/nfs/lab/KG/ref/male.masked.hg19.fa', 
       outdir,  '-bg', bgfile,  '-nomotif' ))
}

mclapply(1:length(ups), function (i) run_homer(paste(beddir,ups[i], sep="/"), 
                                               paste0(hdir,gsub(".bed", "", ups[i]) ), 
                                               paste(beddir,bup, sep="/")),mc.cores = 18)

mclapply(1:length(ups), 
         function (i) run_homer(paste(beddir,downs[i], sep="/"), paste0(hdir,gsub(".bed", "", downs[i]) ), 
                                paste(beddir,bdown, sep="/")),mc.cores = 18)
  

### Plot new motifs with  differential as bckground
up with up / down wth down

homer_dir = hdir
list.files(homer_dir)


files2 = c('cytokine_6hr_Hi.up','cytokine_6hr_Hi.down',
           'cytokine_24hr_Hi2.up','cytokine_24hr_Hi2.down',
           'cytokine_24hr_Hi.up','cytokine_24hr_Hi.down',
            'cytokine_48hr_Hi.up','cytokine_48hr_Hi.down',
            'cytokine_72hr_Hi.up','cytokine_72hr_Hi.down')

pdf("cytokine_figures/Plot_motifs_diffbckg_log.pdf", height = 6, width=14)

par(mfcol=c(2,5), mar = c(4,4,4,0.2))

#par(mfcol=c(2,3), mar = c(4,4,4,0))
for (exp in files2){
hom= read.table( paste0(homer_dir, exp, '/knownResults.txt'), header=F, fill=T, check.names=F, skip=1, stringsAsFactors = F)
hom$fold_change = as.numeric(gsub('%',"", hom$V7)) /(as.numeric(gsub('%',"", hom$V9))+0.1) ##pseudocount for inf (only V2)
hom$Motif = toupper(str_split_fixed(hom$V1, "\\(",5)[,1])

    hom[,"-logP"] = -(as.numeric(hom$V4))
#hom[,"-logP"] = -log10(as.numeric(hom$V3))
 hom = hom[complete.cases(hom),]   
hom=hom[order(hom$fold_change, decreasing=T),]
hom=hom[order(hom[,'-logP'], decreasing=T),]
    
plot(hom[,c('fold_change', '-logP')], pch = 19, 
     col = (hom$V5<0.05)+1  )
if(sum(hom$V5<0.05)>0){
    text (hom[1:25,c('fold_change', '-logP')], hom$Motif[1:20], 
      pos=2, cex=0.7,offset = 0.2)
    
    #text (hom[hom$V5<0.01,c('fold_change', '-logP')], hom$Motif[hom$V5<0.01], 
    #  pos=c(2,4), cex=0.9,offset = 0.5)
}
mtext(exp)
}
dev.off()

### Plot the fold change by time

setwd("/nfs/lab/projects/islet_cytok/analysis/")

library(tidyr)

files3 = c('cytokine_6hr_Hi.up','cytokine_6hr_Hi.down',
           'cytokine_24hr_Hi.up','cytokine_24hr_Hi.down',
            'cytokine_48hr_Hi.up','cytokine_48hr_Hi.down',
            'cytokine_72hr_Hi.up','cytokine_72hr_Hi.down')

homer_dir = '/nfs/lab/elisha/islets/cytokine/HOMER/homer_dec15/'
#list.files(homer_dir)
files3 = c('hi_24_down','hi_24_up','hi2_24_down','hi2_24_up',
           'lo_24_down','lo_24_up','lo2_24_down','lo2_24_up',
           'hi_6_down','hi_6_up','hi_48_down','hi_48_up',
           'hi_72_down','hi_72_up')

df = data.frame()
for (exp in files3){
hom= read.table( paste0(homer_dir, exp, '/knownResults.txt'), header=F, fill=T, check.names=F, skip=1, stringsAsFactors = F)
hom$fold_change = as.numeric(gsub('%',"", hom$V7)) /(as.numeric(gsub('%',"", hom$V9))+0.1) ##pseudocount for inf (only V2)
hom$Motif = toupper(str_split_fixed(hom$V1, "\\(",5)[,1])

hom = hom[complete.cases(hom),] 
hom = subset(hom, Motif !="SEQBIAS:") 
hom$Qval = p.adjust(as.numeric(hom$V3))
tab = hom[,c("Motif","fold_change", "Qval")]
tab = tab[complete.cases(tab),]
tab$treat = exp
    df = rbind(df,tab )
}

head(df)

df = df[!duplicated(df[,c(1,4)]),]
fc = spread(df[, c(1,2,4)], treat, fold_change)
qv = spread(df[, c(1,3,4)], treat, Qval)

rownames(fc) = fc$Motif
fc = fc[,-1]
rownames(qv) = qv$Motif
qv = qv[,-1]

fc = subset(fc, rowSums(qv<0.05, na.rm=T)>0)
qv = subset(qv, rowSums(qv<0.05, na.rm=T)>0)

fc = fc[order(rowSums(log10(qv))),]
qv = qv[order(rowSums(log10(qv))), ]

colnames(fc)

pdf("cytokine_figures/Motifs_by_time_all_bckg.pdf")
par(mfrow = c(5,5), mar =c(3,3,2,2), xpd = T)

#ups   = c(1,3,5,7)
#downs = c(2,4,6,8)

downs   = c(5,1,3,7)
ups = c(6,3,4,8)
for( i in 1:nrow(fc)){
    
    down = data.frame( time = c(6, 24, 48,  72),
                      fold_change = as.numeric(t(fc[i, downs])), qval = as.numeric(t(qv[i, c(1,3,5,7)])) )
    up = data.frame( time = c(6, 24, 48,  72),
                      fold_change = as.numeric(t(fc[i, ups])), qval = as.numeric(t(qv[i, c(2,4,6,8)])) )
    


plot(up[,1:2], type = "o", pch=19, col="orangered",
     axes=F,
     ylim = c(min(c(up[,2], down[,2])), max(c(up[,2], down[,2]))+ max(c(up[,2], down[,2]))/5   )  )
    axis(1, at=c(6, 24, 48,  72), labels = c(6, 24, 48,  72) )
    axis(2)
    mtext("Fold change", side=2 , line=2, cex=0.6)
    abline(h=1, col="gray", lty=2)
text(up[,c(1:2)], c("","*")[(up$qv<0.05) +1], pos=3, cex=2,offset = 0.1)
lines(down[,1:2], type = "o", pch=19, col="dodgerblue")
text(down[,c(1:2)], c("","*")[(down$qv<0.05) +1], pos=3, cex=2,offset = 0.1)
mtext( rownames(fc)[i], cex=0.8)
    }
dev.off()

ranks = apply(-fc, 2 , rank)


head(fc)

#downs   = c(5,11,13,1,9,3,7)
downs   = c(9,5,1,3,7)
downs   = c(5,1,3,7)
ups = downs+1

color = colorRampPalette(brewer.pal(9, "YlGnBu"))(50)

fcup   = fc[,ups]
qup    = qv[,ups]
fcup   = subset(fcup, rowSums(qup<0.05, na.rm=T)>0)
qup    = subset(qup, rowSums(qup<0.05, na.rm=T)>0)

xx = qup
xx[xx<0.05]<-"*"
xx[xx>=0.05]<-""

pheatmap(fcup, col=color, scale="row", cellwidth = 20,display_numbers = xx,
         number_color = "red",
         file = "cytokine_figures/Heatmap_motifs_up_only_hi.pdf",
         cluster_rows = T, cluster_cols = F)

pheatmap(log(fcup,2), col=color, scale="none", cellwidth = 20,display_numbers = xx,
         number_color = "red",
         file = "cytokine_figures/Heatmap_motifs_up_notscales.pdf",
         cluster_rows = T, cluster_cols = F)

fcdo   = fc[,downs]
qdo    = qv[,downs]
fcdo   = subset(fcdo, rowSums(qdo<0.05, na.rm=T)>0)
qdo    = subset(qdo, rowSums(qdo<0.05, na.rm=T)>0)

xx = qdo
xx[xx<0.05]<-"*"
xx[xx>=0.05]<-""

pheatmap(fcdo, col=color, scale="row", cellwidth = 20,display_numbers = xx,
         number_color = "red",
         file = "cytokine_figures/Heatmap_motifs_down_onlyhi.pdf",
         cluster_rows = T, cluster_cols = F)

l2 = log(fcdo,2)
l2[l2==-Inf]<-0

pheatmap(l2, col=color, scale="none", cellwidth = 20,display_numbers = xx,
         number_color = "red",
         file = "cytokine_figures/Heatmap_motifs_down_log2fc.pdf",
         cluster_rows = T, cluster_cols = F)

### Supplementary Table

homer_dir = '/nfs/lab/elisha/islets/cytokine/HOMER/homer_dec15/'
list.files(homer_dir)

files3 = c('union_down', 'union_up',
           'hi_6_down','hi_6_up',
           'hi_24_down','hi_24_up',
           'hi_48_down','hi_48_up',
           'hi_72_down','hi_72_up' ,
           'lo_24_down','lo_24_up',
           'lo2_24_down','lo2_24_up',
           'hi2_24_down','hi2_24_up')

cn = c('Motif','P_value', "log_P_value",'Q_value','No_seq_withMotif','Pct_seq_withMotif','Fold_change')

for (u in 0:1){
    for(s in seq(1,16,by=2)+u){
exp =     files3[s]
hom= read.table( paste0(homer_dir, exp, '/knownResults.txt'), header=F, fill=T, check.names=F, skip=1, stringsAsFactors = F)
hom$fold_change = as.numeric(gsub('%',"", hom$V7)) /(as.numeric(gsub('%',"", hom$V9))+0.1) ##pseudocount for inf (only V2)
hom$Motif2 = toupper(str_split_fixed(hom$V1, "\\(",5)[,1])
hom$Motif = toupper(str_split_fixed(hom$V1, "\\/",3)[,1])
hom = hom[complete.cases(hom),] 
hom = subset(hom, Motif2 !="SEQBIAS:") 
    
tab = hom[,c(12,3,4, 5,6,7,10)]
colnames(tab) = paste(cn, s , sep=".")
tab$Rank = 1:nrow(tab)
   if( s == (1+u) ) 
   {df =tab }else{
       df = merge(df, tab, by="Rank", all.x=T)
   }
}
    if(u==0){df_down=df}
      }

colnames(df) = colnames(df_down)=c("Rank", rep(cn, 8))

write.csv(df_down, "HOMER_diff_sites/SuppTab_down.csv", row.names=F )

write.csv(df, "HOMER_diff_sites/SuppTab_up.csv", row.names=F )

### GeoTable

mat = read.table('/nfs/lab/elisha/islets/cytokine/update_deseq/matrix/cytPeak.dec9.matrix.filtered.R', header=T, row.names=1)

meta = read.table('/nfs/lab/elisha/islets/cytokine/update_deseq/metadata/cytokine_35.metadata.txt', header=T)

meta$recode = meta$donor

meta$recode[meta$donor=="islet_1"]<-'SAMN12833535'
meta$recode[meta$donor=="islet_2"]<-'SAMN12889245'

table(meta$condition)

dim(mat)

recode = c(hi2='2cyt-hi',
lo2='2cyt-lo',
hi='3cyt-hi',
lo='3cyt-lo',
untreated = 'untreat')

meta$treat = recode[meta$condition]

meta$sampname = paste(meta$recode, meta$treat, paste0(meta$time, "h"), sep="_")

colnames(mat) = meta$sampname[match( colnames(mat), meta$bam)]

rownames(mat) = gsub("peak-", "", rownames(mat))

write.table(mat, "../geo_submission_oct21/processed_files/Bulk_ATACSeq_count.matrix.txt", sep="\t", quote=F)

meta[meta$recode=='SAMN10977276',]
