==> notebooks/06_BulkATAC_uploadBigWig.ipynb <==
library(stringr)
library(RColorBrewer)

#### This takes all .bw files that are in the speficied directory 

bwdir = '/nfs/lab/projects/islet_cytok/data/ATACseq/bigWig_bulkATAC_merged'

bws = list.files(bwdir)[grepl(".bw",list.files(bwdir))]

samples = str_remove (bws, "_merge.bw")

samples

#### Copy the .bw files in the amazon bucket

awsdir = 's3://gaulton-lab-ucsc/treated_islet_new/'

commands = paste0("aws s3 cp " , bwdir,"/", bws," " ,awsdir , bws)

sapply (commands, system)

#### Make track lines to inster tin the ucsc browser (copy-paste)

color = brewer.pal(5, "Set1")[c(5,1,1,1,1,2,3,4,4,4,4)]

rgb_colors   = sapply (color , function(x) paste(col2rgb(x), collapse=","))

for( i in 1:length(samples)) {
nam    = samples[i]
col    = rgb_colors [i]
url    = paste0('https://s3-us-west-2.amazonaws.com/gaulton-lab-ucsc/treated_islet_new/', bws[i])
text   = paste("track name=\"", nam, "\" color=", col, " type=bigWig bigDataUrl=", url, 
               " visibility=2 maxHeightPixels=20 smoothingWindow=2 autoScale=off viewLimits=0:30 yLineOnOff=on", sep="")
cat(text, "\n")    
    }
