==> notebooks/12_snATAC_peakcalls_annotate.ipynb <==
Here are the peak calls for the snATAC-seq data (excluding biobank samples) by cell type:
- cytokine+untreated:/nfs/lab/elisha/islets/cytokine/new_islet/gaulton/peaks/gatsby/cellType_cytunt
- cytokine: /nfs/lab/elisha/islets/cytokine/new_islet/gaulton/peaks/gatsby/cellType_cyt
- untreated: /nfs/lab/elisha/islets/cytokine/new_islet/gaulton/peaks/gatsby/cellType_unt 

library(stringr)
library(RColorBrewer)

setwd("/nfs/lab/projects/islet_cytok/analysis/snATAC/peaks")

annp = read.table('Islet_bulkATAC.subpop.sorted.merged.bed')

annp$counts = str_count(annp$V4,",")+1
annp$peaks = paste("peak", annp[,1], annp[,2], annp[,3], sep="-")

u = read.table("../../deseq_ATAC_NEW_bedfiles/annotated_peaks.tsv")

annp$upr = annp$peaks %in% rownames(u) [rowSums(u[,c(2,4,6,8,10,12,14)])>0]
annp$dor = annp$peaks %in% rownames(u) [rowSums(u[,c(1,3,5,7,9,11,13)])>0]

sum(rowSums(u)>0)

head(annp)

bed1 = read.table('Islet_bulkATAC.bed' )
bed2 = read.table('Islet_bulkATAC.subpop.sorted.merged.bed')

bed1$V4="z"
bedm = rbind(bed1, bed2)

bedm = bedm[order(bedm$V4),]

bedm = bedm[!duplicated(bedm[,1:3]),]

bedm$V4[bedm$V4=='z']<-"na"

peaks = paste("peak", bedm[,1], bedm[,2], bedm[,3], sep="-")

sum(duplicated(annp$peaks))

dim(bedm)

bedm$cyt = peaks %in% rownames(u)

tab1 = subset(bedm, grepl("beta", bedm$V4))

colnames(bedm) = c("chrom", "start", "end", "islet_cell_type", "cytokine_responsive")

write.table(bedm, "cCREs_annotated.bed", sep="\t", quote=F, row.names=F)

cells = c('acinar','alpha','beta','delta','endothelial','stellate')
mat1 =  matrix(NA, nrow=2, ncol=6)
colnames(mat1) = cells
rownames(mat1) = c("down", "up")
mat2 = mat1
for (cell in cells ){
    tab1 = subset(annp, grepl(cell, annp$V4))
    write.table(tab1[tab1$upr==T,1:3], paste0("/nfs/lab/projects/islet_cytok/analysis/ldsc/", cell, ".up.bed"), 
                sep="\t", quote=F, row.names=F, col.names=F)
    write.table(tab1[tab1$dor==T,1:3], paste0("/nfs/lab/projects/islet_cytok/analysis/ldsc/", cell, ".down.bed"), 
                sep="\t", quote=F, row.names=F, col.names=F)
    mat1["up",cell] = sum(tab1$upr)
    mat1["down",cell] = sum(tab1$dor)
    
    tab2 = subset(annp, annp$V4 == cell)
    mat2["up",cell] = sum(tab2$upr)
    mat2["down",cell] = sum(tab2$dor)
}

mat1
colSums(mat1)

3186/38978

mat2
colSums(mat2)

pdf("../../cytokine_figures/snatac_annotated_peaks.pdf")
par(mfrow=c(2,2),mar = c(4,4,4,5))
b = barplot(mat1, las=3,
           ylim=c(0,50000), col = brewer.pal(9, "PuOr")[c(3,7)] )

text(x = b, y =mat1[1,], labels = paste0(round(unlist(mat1[1 ,])/1000,1),"k") , pos = 1, cex=0.75)
text(x = b, y = as.numeric(mat1[2,])+ as.numeric(mat1[1,]), 
     labels = paste0(round(unlist(mat1[2,])/1000,1),"k") , pos = 1, cex=0.75)

b = barplot(mat2, las=3,
           ylim=c(0,3500), col = brewer.pal(9, "PuOr")[c(3,7)] )

text(x = b, y =mat2[1,], labels = paste0(round(unlist(mat2[1 ,])/1000,1),"k") , pos = 1, cex=0.75)
text(x = b, y = as.numeric(mat2[2,])+ as.numeric(mat2[1,]), 
     labels = paste0(round(unlist(mat2[2,])/1000,1),"k") , pos = 1, cex=0.75)
dev.off()

 tab2 = subset(annp, V4 == "beta" & upr == TRUE)

head(tab2)

cyt = read.table('/nfs/lab/elisha/islets/cytokine/update_deseq/output/cytokine_24hr_Hi.dec16.filtered.txt')
cyt = cyt[order(cyt$pvalue),]

cytb = subset(cyt, rownames(cyt) %in% tab2$peaks)

head(cytb,20)

head(annp)

## Motif enrichment in beta cell specific peaks 

tab1 = subset(annp, V4=="beta")
    write.table(tab1[tab1$upr==T,1:3],  "homer_beta/beta_spec.up.bed", 
                sep="\t", quote=F, row.names=F, col.names=F)
    write.table(tab1[tab1$dor==T,1:3],  "homer_beta/beta_spec.down.bed", 
                sep="\t", quote=F, row.names=F, col.names=F)

write.table(annp[annp$upr==T,1:3],  "homer_beta/all.up.bed", 
                sep="\t", quote=F, row.names=F, col.names=F)
    write.table(annp[annp$dor==T,1:3],  "homer_beta/all.down.bed", 
                sep="\t", quote=F, row.names=F, col.names=F)

run_homer = function(bedfile, outdir, bgfile )
{
    system(paste ('findMotifsGenome.pl', bedfile, '/nfs/lab/KG/ref/male.masked.hg19.fa', 
       outdir,  '-bg', bgfile,  '-nomotif' ))
}

run_homer(bedfile = paste0(getwd(),"/homer_beta/beta_spec.up.bed"),
          outdir  = paste0(getwd(),"/homer_beta/beta_spec.up"),
          bgfile  = paste0(getwd(),"/homer_beta/all.up.bed"))

run_homer(bedfile = paste0(getwd(),"/homer_beta/beta_spec.down.bed"),
          outdir  = paste0(getwd(),"/homer_beta/beta_spec.down"),
          bgfile  = paste0(getwd(),"/homer_beta/all.down.bed"))
