==> notebooks/19_SNP_SELEX_MotifBreakR.ipynb <==
suppressPackageStartupMessages(library(motifbreakR))
suppressPackageStartupMessages(library(MotifDb))
suppressPackageStartupMessages(library(BSgenome))
suppressPackageStartupMessages(library(tidyr))
suppressPackageStartupMessages(library("BSgenome.Hsapiens.UCSC.hg19"))
suppressPackageStartupMessages(library(motifStack))
suppressPackageStartupMessages(library(data.table))
suppressPackageStartupMessages(library(stringr))
suppressPackageStartupMessages(library(RColorBrewer))
suppressPackageStartupMessages(library(gridExtra))
suppressPackageStartupMessages(library(ggplot2))
suppressPackageStartupMessages(library(MotIV))

setwd('/nfs/lab/projects/islet_cytok/analysis/selex/motifbreak')

vars = read.table("../selex_variants_reclassified.bed")

coord = str_replace_all(vars$V4,"\\_",":")
coord = gsub("T1D:", "", coord)
coord = gsub("T2D:", "", coord)
varlist= vars[,1:3]
varlist[,4] = paste0("chr", coord)
varlist[,5]= 0
varlist[,6]="+"
varlist = varlist[!duplicated(varlist),]

write.table( varlist, "Input_var_list.tsv", sep="\t", quote=F, row.names=F, col.names=F)

dim(varlist)

## Test analysis

hocomoco2 = MotifDb[mcols(MotifDb)$dataSource=='HOCOMOCOv10' & mcols(MotifDb)$organism=='Hsapiens']

snps <- snps.from.file(file = "Input_var_list.tsv",
                                  search.genome = BSgenome.Hsapiens.UCSC.hg19,
                                  format = "bed")

results <- motifbreakR(snpList = snps[1], filterp = TRUE,
                       pwmList = hocomoco2,
                       method="ic",
                       threshold = 5e-4,
                       BPPARAM = BiocParallel::bpparam("SerialParam"))


length(snps)

181540 + 402 +401

length(181401:181540)

#### Run using /bin/Rscript Motifbreak.R (~16 hrs) -- otherwise will load anaconda3 R
# for (i in seq(1,181400, by=200)){
#   test=snps[i:(i+199)]
#   results <- mclapply(1:200, function(x) data.frame(motifbreakR(snpList = test[x], filterp = TRUE,
#                        pwmList = hocomoco2,
#                        method="ic",
#                        threshold = 5e-4,
#                        BPPARAM = BiocParallel::bpparam("SerialParam"))), mc.cores=18 )

# rb = rbindlist(results[lapply(results, class) == "data.frame"])

# if(nrow(rb)>0){
#   write.table(rb, paste0("hocomoco_results/Results_Hocomoco_", i, "_", i+199, ".tsv"), quote=F, row.names=F, sep="\t")  
# }

# }

#   test=snps[181401:181540]
#   results <- mclapply(1:140, function(x) data.frame(motifbreakR(snpList = test[x], filterp = TRUE,
#                        pwmList = hocomoco2,
#                        method="ic",
#                        threshold = 5e-4,
#                        BPPARAM = BiocParallel::bpparam("SerialParam"))), mc.cores=18 )

# rb = rbindlist(results[lapply(results, class) == "data.frame"])

# if(nrow(rb)>0){
#   write.table(rb, paste0("hocomoco_results/Results_Hocomoco_", 181401, "_", 181540, ".tsv"), quote=F, row.names=F, sep="\t")  
# }


files = list.files('hocomoco_results')[grepl("Results", list.files('hocomoco_results'))]

length(files)

RES = data.frame()
for( f in files){
    res = read.table(paste0('hocomoco_results/',f),  row.names=NULL, header=T, sep="\t", stringsAsFactors = F)
    RES = rbind(RES, res)
}

tab = RES[, c(6,10:24)]
tab = tab[order(-abs(tab$alleleDiff)),]
tab = tab[!duplicated(tab[,c('SNP_id','geneSymbol' )]),]

tab[tab$SNP_id=="chr14:68836369:A:G",]

write.table(tab, "Summary_significant_motifs_hocomocov10_long.tsv", sep="\t", row.names=F, quote=F)

min(abs(tab$alleleDiff))

dim(tab)

sum(!duplicated(tab$SNP_id))

selex = read.table("../combinedResults.tsv", header=T)

selex = subset(selex, !is.na(PBSb))
coord = str_replace_all(selex$snp_name,"\\_",":")
coord = gsub("T1D:", "", coord)
coord = gsub("T2D:", "", coord)
selex$snpID = paste0 ("chr", coord)

comp = merge(selex[,c('snpID','prot', "Family", 'PBSb', 'PBSb_pv')], tab[,c('SNP_id', 'geneSymbol','alleleDiff') ],
             by=1:2, all.x=T)

dim(comp)

sum(!duplicated(comp$prot))

sum(!duplicated(comp$snpID))

citation("motifbreakR")

length(hocomoco2)
