==> notebooks/21_SOCS1_figure.ipynb <==
setwd('/nfs/lab/projects/islet_cytok/analysis/socs1_figure')

library(scales)
library(RColorBrewer)
suppressPackageStartupMessages(library(plot3D))

rec = read.csv("../selex/selex_variants_reclassified_annotated_with_atac_new.csv", row.names=1)

t1d = read.table('../selex/T1D.99credset_new.bed', stringsAsFactors = F)

finemap = subset(t1d, grepl("DEXI", t1d$V5))

finemap$pbsnp_diffatac_1 = finemap$V4 %in% rec$snpID[rec$pbs005 & rec$diff==T]

finemap$pbsnp_atac_1 = finemap$V4 %in% rec$snpID[rec$pbs005 & rec$atac==T]

finemap$pbsnp = finemap$V4 %in% rec$snpID[rec$pbs005 ]

table(finemap$pbsnp_diffatac_1,finemap$pbsnp_atac_1)

table(finemap$pbsnp)

# bedtools intersect -a /nfs/lab/publicdata/1kg_snps/ALL.wgs.phase3_sites.bed.ids -b loci -wa > loci_snps.bed
# awk -v OFS="\t" '{print $5}' loci_snps.bed > loci_snps.rsid_list
# zgrep -F -f loci_snps.rsid_list /home/joshchiou/joshchiou-data2/T1D-GWAS/matched_cohorts/TOPMed/meta_analysis/T1D_meta.all_chr.hg19.sumstats.txt.gz > T1D.Chiou2020.select

gwas = read.table("T1D.Chiou2020.select", sep="\t", stringsAsFactors = F)
gwas$log10pval = -log10(gwas$V8)
gwas$credset = gwas$V1 %in% t1d$V4

ld = read.table("caqtl.ld.EUR.ld", header=T)

ld2 = read.table("caqtl.ld.EUR2.ld", header=T)

dim(ld)
dim(ld2)
dim(gwas)

gwas = merge(gwas, ld[,6:7], by=1, all.x=T)

gwas$R2[gwas$V1 == '16:11194771:C:T']<-1
gwas$R2[is.na(gwas$R2)]<- 0

gwas = merge(gwas, ld2[,6:7], by=1, all.x=T)

gwas$R2.y[gwas$V1 == '11428643:C:G']<-1
gwas$R2.y[is.na(gwas$R2.y)]<- 0

my_palette = colorRampPalette( brewer.pal(9, "Blues")[3:9] )(20)
my_palette2 = colorRampPalette( brewer.pal(9, "Reds")[3:9] )(20)

color1   <- my_palette[as.numeric(cut(gwas$R2.x ,breaks = 20))]
color2  <- my_palette2[as.numeric(cut(gwas$R2.y ,breaks = 20))]

color = c(color1[1:head(which(gwas$V3 >11350000 ),1)],
          color2[head(which(gwas$V3 >11350000 ))[2]:length(color2)])

gwas$color = color
gwas$color[gwas$R2.y == 0 & gwas$R2.x==0]<-"gray"
gwas_toplot = gwas[order(gwas$log10pval),]

gwas_toplot = gwas_toplot[order(gwas_toplot$R2.x, decreasing = F),]
gwas_toplot = gwas_toplot[order(gwas_toplot$R2.y, decreasing = F),]

pdf("socs1_finemap.pdf")
par(mfrow=c(3,1), mar = c(4,2,6,4), las=1)
plot(log10pval~V3, gwas_toplot,  xlim= c(10934642,11913586), ylim=c(0,45),
     pch= 21,
     col = c(NA, alpha('white',0.5))[gwas_toplot$credset +1],
     #col=alpha('gray', 0.5),
     bg=alpha(gwas_toplot$color, 0.8)
     #col=alpha(c("gray", "red"),0.5)[gwas$credset +1]
     )
colkey(c("gray",my_palette), clim=c(0,1), clab="R2", add=T, side=4, length=0.5)
# plot(V7~V3, finemap, 
#   col=alpha(c("black", "red"),0.6)[finemap$pbsnp_diffatac_1 +1],
#      pch=16, xlim= c(10934642,11913586))

plot(V7~V3, finemap[finemap$V5 == 'DEXI_16:11194771:C:T',], 
     col=alpha(c("black", "red"),0.8)[finemap$pbsnp_diffatac_1[finemap$V5 == 'DEXI_16:11194771:C:T'] +1],
     pch=16, xlim= c(10934642,11913586))
 
plot(V7~V3, finemap[finemap$V5 == 'DEXI_16:11428643:C:G',], 
      col=alpha(c("black", "red"),0.8)[finemap$pbsnp_diffatac_1[finemap$V5 == 'DEXI_16:11428643:C:G'] +1],
   pch=16, xlim= c(10934642,11913586))

colkey(c("gray",my_palette2), clim=c(0,1), clab="R2", add=T, side=4, length=0.5)

    dev.off()


png("socs1_locus.png", width = 6.51, height = 1.064, units = 'in', res = 300)
par( mar=c(0,0,0,0))
plot(log10pval~V3, gwas_toplot,  xlim= c(10934642,11913586), ylim=c(0,45),
     pch= 21,axes=F,lwd=0.5, cex=0.6,
     col = c(NA, alpha('white',0.5))[gwas_toplot$credset +1],
     #col=alpha('gray', 0.5),
     bg=alpha(gwas_toplot$color, 0.8)
     #col=alpha(c("gray", "red"),0.5)[gwas$credset +1]
     )
dev.off()

png("socs1_locus_norbord.png", width = 6.51, height = 1.064, units = 'in', res = 300)
par( mar=c(0,0,0,0))
plot(log10pval~V3, gwas_toplot,  xlim= c(10934642,11913586), ylim=c(0,45),
     pch= 19,axes=F,cex=0.6,
     col=alpha(gwas_toplot$color, 0.8)
     
     )
dev.off()

write.table(finemap[finemap$V5 == 'DEXI_16:11428643:C:G',1:3], "finemap_snps.bed",
           sep="\t", col.names=F, row.names=F, quote=F)

## APOTRACKER experiment

suppressPackageStartupMessages(library(stringr))
suppressPackageStartupMessages(library(RColorBrewer))
suppressPackageStartupMessages(library(gplots))
suppressPackageStartupMessages(library(beeswarm))
suppressPackageStartupMessages(library(tidyr))

new = read.table("endoc_ko.txt", row.names=1)

df = data.frame(Treatment = c(rep('Control', 8), rep('Cytokine', 8)), 
                shRNA = rep(c(rep('Scramble', 4), rep('SOCS1', 4)),2 ) ,
          meas = c(t(new[1,]), t(new[2,])))

df

anova(lm(meas ~ shRNA , df[df$Treatment=='Control',]))

anova(lm(meas ~ shRNA , df[df$Treatment=='Cytokine',]))

res.aov = aov(meas ~ shRNA*Treatment , df)

summary(res.aov)

TukeyHSD(res.aov)
